backend/rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.2
firebase-admin==6.2.0
python-dotenv==1.0.0
numpy>=1.26.0
PyPDF2==3.0.1
requests==2.31.0
rank-bm25==0.2.2
scikit-learn==1.3.0
google-cloud-aiplatform-v1==1.38.1
genkit==0.1.0
unstructured[pdf]==0.12.0
unstructured-inference==0.7.0 